remove usage view caching (idea merge)
Get rid of UsageViewImpl renderer cache optimization  It was a workaround for performance problems in VariableHeightLayoutCache, but this workaround itself had some performance problems. Now that we've replaced VariableHeightLayoutCache with DefaultTreeLayoutCache that doesn't try to access renderers for nodes outside the visible rectangle, this workaround is no longer necessary and does more harm than good.  It appears that it was also queuing updates on the visible nodes in the process of clearing the renderer cache, but tree updates seem to work just fine without it. If it turns out there's a scenario when tree update doesn't work anymore, it should be fixed separately, by updating nodes in background, like it should be done.